pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "datathon-grupo-xx"
version = "0.1.0"
description = "Datathon FIAP Fase 05 - Plataforma MLOps para previsao de series temporais"
readme = "README.md"
requires-python = ">=3.11,<3.13"
authors = [{ name = "Lima", email = "lima@example.com" }]
license = { text = "MIT" }

dependencies = [
    # ---- Core ML ----
    # IMPORTANTE: TensorFlow 2.16.x ja vem com Keras 3 bundled.
    # NAO instalar `keras` separadamente - causa conflito de versoes.
    "numpy>=1.26,<2.0",
    "pandas>=2.2",
    "scikit-learn>=1.5",
    "tensorflow==2.16.2",
    "joblib>=1.3",

    # ---- Ingestao de dados ----
    "yfinance>=0.2.40",
    "pyarrow>=15.0",

    # ---- Tracking e versionamento ----
    "mlflow>=2.14",
    "dvc>=3.50",

    # ---- Serving ----
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.5",
    "python-multipart>=0.0.9",

    # ---- Observabilidade (Gap 01) ----
    "prometheus-fastapi-instrumentator>=7.0",
    "prometheus-client>=0.20",

    # ---- Drift e validacao (Gap 04, 06) ----
    "evidently>=0.4.30",
    "pandera>=0.18",

    # ---- Config ----
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.5",
    "mypy>=1.10",
    "bandit[toml]>=1.7",
    "pre-commit>=3.7",
    "types-PyYAML",
    "types-requests",
]

[project.urls]
Repository = "https://github.com/grupo-xx/datathon-fase05"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
exclude = ["tests*", "notebooks*", "docs*"]

# ============================================================================
# Ruff
# ============================================================================
[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]
extend-exclude = ["notebooks", ".venv", "build", "dist", "_legacy"]

[tool.ruff.lint]
select = [
    "E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "D",
]
ignore = [
    "D100", "D104", "D107",
    "D203", "D213",
    "E501",
    "N806",   # X maiusculo - convencao em ML para matrizes de features
    "B008",   # FastAPI: Depends/File em defaults sao idiomaticos
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D"]
"notebooks/**" = ["D", "E", "F"]
"src/**/__init__.py" = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

# ============================================================================
# mypy
# ============================================================================
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = false
check_untyped_defs = true
exclude = ["notebooks", "build", "dist", "_legacy"]

# ============================================================================
# pytest
# ============================================================================
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = ["-v", "--strict-markers", "--tb=short"]
markers = [
    "slow: testes que demoram mais de 5s",
    "integration: testes que dependem de modelo treinado, rede ou docker",
]

# ============================================================================
# coverage
# ============================================================================
[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__init__.py"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
show_missing = true
skip_covered = false

# ============================================================================
# bandit
# ============================================================================
[tool.bandit]
exclude_dirs = ["tests", "notebooks", ".venv", "_legacy"]
skips = ["B101"]
